requirements.txt
appnope==0.1.0
attrs==19.1.0
backcall==0.1.0
bleach==3.1.0
cachetools==4.1.0
certifi==2019.3.9
chardet==3.0.4
cycler==0.10.0
decorator==4.4.0
defusedxml==0.6.0
docopt==0.6.2
entrypoints==0.3
google-api-core==1.17.0
google-api-python-client==1.8.0
google-auth==1.14.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
googleapis-common-protos==1.51.0
httplib2==0.17.2
idna==2.9
ipykernel==5.1.1
ipython==7.5.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
jedi==0.13.3
Jinja2==2.10.1
joblib==0.13.2
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.1.0
mistune==0.8.4
motivate==0.0.1
nbconvert==5.5.0
nbformat==4.4.0
notebook==5.7.8
numpy==1.16.4
oauthlib==3.1.0
pandas==0.24.2
pandocfilters==1.4.2
parso==0.4.0
pexpect==4.7.0
pickleshare==0.7.5
prettytable==0.7.2
prometheus-client==0.7.0
prompt-toolkit==2.0.9
protobuf==3.11.3
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.4.2
pyparsing==2.4.0
pyrsistent==0.15.2
PySimpleGUI==4.14.0
python-dateutil==2.8.0
pytz==2019.1
PyYAML==5.3.1
pyzmq==18.0.1
qtconsole==4.5.1
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
scikit-learn==0.21.2
scipy==1.3.0
Send2Trash==1.5.0
six==1.12.0
spotify==0.0.1
terminado==0.8.2
testpath==0.4.2
tornado==6.0.2
traitlets==4.3.2
uritemplate==3.0.1
urllib3==1.25.9
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.4.2
